moved former ElboDeriv functions to log_prob_util.jl, cleaned up imports
